+++ mcww_web/js/misc/fileOpen.js
@@ -13,18 +13,14 @@ if ("launchQueue" in window) {
             } catch (error) {
                 console.log(error);
             }
+            document.querySelector(".file-open-buttons-new-window").classList.remove("mcww-hidden");
+            document.querySelector(".file-open-buttons-same-window").classList.add("mcww-hidden");
             const handledFile = launchParams.files[0];
             const blob = await handledFile.getFile();
             const url = window.URL.createObjectURL(blob);
             copyMediaToClipboard(url, handledFile.name);
             waitForElement(document, ".opened-file button.paste", (button) => {
                 button.click();
-                waitForElement(document, ".opened-file .download-link", (link) => {
-                    if (!isModel3DUrl(link.href)) {
-                        copyMediaToClipboard(link.href);
-                        grInfo("Opened file has been copied into clipboard");
-                    }
-                });
                 setSessionStorageVariable("fileOpenHandled", true);
             });
         } else {
@@ -36,7 +32,7 @@ if ("launchQueue" in window) {
 }
 
 
-function updateOpenedFileNameInTitle() {
+function afterFileOpened() {
     const fileStem = document.querySelector(".opened-file .filename .stem")?.textContent;
     const fileExt = document.querySelector(".opened-file .filename .ext")?.textContent;
     let fileName = null;
@@ -44,4 +40,37 @@ function updateOpenedFileNameInTitle() {
         fileName = fileStem + fileExt;
     }
     TITLE.setOpenedFileName(fileName);
+    if (fileName) {
+        waitForElement(document, ".opened-file .download-link", (link) => {
+            if (!isModel3DUrl(link.href)) {
+                copyMediaToClipboard(link.href);
+                grInfo("Opened file has been copied into clipboard");
+            }
+        });
+    }
 }
+
+
+document.addEventListener('drop', (event) => {
+    event.preventDefault();
+    const files = [...event.dataTransfer.items]
+    .map((item) => item.getAsFile())
+    .filter((file) => file);
+    if (files.length > 1) {
+        grError("Only 1 file drop is supported");
+    }
+    if (files.length === 1) {
+        if (getSelectedMainUIPage() !== "fileOpen") {
+            selectMainUIPage("fileOpen");
+        }
+        document.querySelector(".file-open-buttons-new-window").classList.add("mcww-hidden");
+        document.querySelector(".file-open-buttons-same-window").classList.remove("mcww-hidden");
+        const file = files[0];
+        const blob = new Blob([file], { type: file.type });
+        const url = window.URL.createObjectURL(blob);
+        copyMediaToClipboard(url, file.name);
+        waitForElement(document, ".opened-file button.paste", (button) => {
+            button.click();
+        });
+    }
+});