+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### (WIP) 2.2
 - PWA files association with 3D models, so you can use MCWW as a simple 3D models viewer. It doesn't support nodes with 3D models in workflows yet. If you installed PWA before, you need to reinstall it to register these extensions associations
+- Can open file on drop anywhere in the document
 - Split presets dataset in pages for performance reasons. If there are too many presets, workflow rendering takes too much time. Default page size is 50, can be changed in the settings
 
 ### 2.1 - Overflow galleries & Markdown update